rework sceneloader and cache
