+++ Engine/src/Mesh/MeshInfo.h
@@ -11,8 +11,8 @@ struct MeshInfo {
   MeshInfo() = delete;
   MeshInfo(unsigned int a, unsigned int elcount);
   MeshInfo(unsigned int a, unsigned int elcount, glm::mat4 trans);
-  MeshInfo(unsigned int a, unsigned int elcount, TextureMapType t_id, glm::mat4 trans);
-  MeshInfo(unsigned int a, unsigned int elcount, TextureMapType t_id, Material mat, glm::mat4 trans);
+  MeshInfo(unsigned int a, unsigned int elcount, TextureMapType t_id, glm::mat4 local_trans);
+  MeshInfo(unsigned int a, unsigned int elcount, TextureMapType t_id, Material mat, glm::mat4 local_trans);
   unsigned int vao = 0;
   unsigned int numElements = 0;
   TextureMapType textureDrawIds{}; // map of <draw_id, details>
