+++ Engine/src/OS/OpenGL/PrimativeMaker.h
@@ -1,8 +1,10 @@
+// only OpenGL class can use this extension
 #pragma once
 namespace AA {
 class PrimativeMaker {
-public:
-  static unsigned int load_cone(unsigned int& out_num_elements);
+private:
+  friend class OpenGL;
+  static unsigned int load_cone();
   static void unload_cone();
   static unsigned int load_cube();
   static void unload_cube();
