+++ UnitTest/UnitTest.cpp
@@ -220,9 +220,9 @@ public:
     }
 
     // load models
-    tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(-20, 0, -25));
-    tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->fullgroundplane.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(2));
-    tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->fullpeasant_man.c_str(), false, glm::vec3(0, -30, -100), glm::vec3(.25f));
+    tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(-20, 0, -25));
+    tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->ground_plane_runtime_dir_path.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(2));
+    tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->peasant_man_runtime_dir_path.c_str(), false, glm::vec3(0, -30, -100), glm::vec3(.25f));
 
     tg->g_update_func = tg->g_aa_interface.AddToUpdate([](float dt) {
       static float accum_time = 0;
@@ -280,9 +280,16 @@ public:
       setup_fpp_fly(tg->g_cam_id, tg->g_aa_interface);
     }
 
-    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->fullzombie_.c_str(), true, glm::vec3(0, -30, -75), glm::vec3(.25f));
-    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->fullzombie_.c_str(), tg->g_zombie_id[0]);
-    tg->g_aa_interface.SetAnimationOnAnimProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
+    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->zombie_runtime_dir_path.c_str(), true, glm::vec3(0, -30, -75), glm::vec3(.25f));
+    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->zombie_runtime_dir_path.c_str(), tg->g_zombie_id[0]);
+    tg->g_aa_interface.SetAnimationOnProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
+    
+    // turn off backfacing culling (because cloth)
+    //{
+    //  auto zombieref = tg->g_aa_interface.GetProp(tg->g_zombie_id[0]).lock();
+    //  zombieref->SetBackfaceCull(false);
+    //}
+
 
     // load sun
     load_sun(tg->g_aa_interface);
@@ -347,7 +354,7 @@ public:
     }
 
     {
-      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(-10, -10, -10));
+      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(-10, -10, -10));
       auto w1 = tg->g_aa_interface.GetProp(tg->g_untextured_cube_id[0]);
       std::shared_ptr<AA::Scene> s1 = w1.lock();
       s1->SetRotation(glm::vec3(1.571f, 3.14159f, 0));
@@ -357,17 +364,17 @@ public:
       s1->SetStencilScale(1.1f);
     }
 
-    tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->fullgroundplane.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(3));
+    tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->ground_plane_runtime_dir_path.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(3));
 
 
-    //tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->fullpeasant_man.c_str(), glm::vec3(0, -30, -70), glm::vec3(.15f));
+    //tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->peasant_man_runtime_dir_path.c_str(), glm::vec3(0, -30, -70), glm::vec3(.15f));
 
 
 
     // Add Zombie With Punching Animation.
-    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->fullzombie_.c_str(), true, glm::vec3(-30, -30, -70), glm::vec3(0.12f));
-    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->fullzombie_.c_str(), tg->g_zombie_id[0]);
-    tg->g_aa_interface.SetAnimationOnAnimProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
+    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->zombie_runtime_dir_path.c_str(), true, glm::vec3(-30, -30, -70), glm::vec3(0.12f));
+    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->zombie_runtime_dir_path.c_str(), tg->g_zombie_id[0]);
+    tg->g_aa_interface.SetAnimationOnProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
     {
       auto weak_tmp = tg->g_aa_interface.GetProp(tg->g_zombie_id[0]);
       auto strong_tmp = weak_tmp.lock();
@@ -431,7 +438,7 @@ public:
 
     // load models
     {
-      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(0, 0, -10));
+      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(0, 0, -10));
       auto w1 = tg->g_aa_interface.GetProp(tg->g_untextured_cube_id[0]);
       std::shared_ptr<AA::Scene> s1 = w1.lock();
       s1->SetStencil(true);
@@ -439,12 +446,12 @@ public:
       s1->SetStencilWithNormals(false);
       s1->SetStencilScale(1.1f);
 
-      tg->g_untextured_cube_id[1] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(-10, 0, -10));
-      tg->g_untextured_cube_id[2] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(10, 0, -10));
+      tg->g_untextured_cube_id[1] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(-10, 0, -10));
+      tg->g_untextured_cube_id[2] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(10, 0, -10));
 
-      tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->fullgroundplane.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(3, 1, 3));
+      tg->g_ground_plane_id = tg->g_aa_interface.AddProp(tg->ground_plane_runtime_dir_path.c_str(), false, glm::vec3(0, -30.f, 0), glm::vec3(3, 1, 3));
 
-      tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->fullpeasant_man.c_str(), false, glm::vec3(0, -30, -100), glm::vec3(.25f));
+      tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->peasant_man_runtime_dir_path.c_str(), false, glm::vec3(0, -30, -100), glm::vec3(.25f));
       auto w2 = tg->g_aa_interface.GetProp(tg->g_peasant_man_id);
       std::shared_ptr<AA::Scene> s2 = w2.lock();
       s2->SetStencil(true);
@@ -452,7 +459,7 @@ public:
       s2->SetStencilWithNormals(true);
       s2->SetStencilScale(3.f);
 
-      tg->g_walking_man_id = tg->g_aa_interface.AddProp(tg->fullwalking_man.c_str(), false, glm::vec3(-60, -30, -100), glm::vec3(.25f));
+      tg->g_walking_man_id = tg->g_aa_interface.AddProp(tg->peasant_girl_runtime_dir_path.c_str(), false, glm::vec3(-60, -30, -100), glm::vec3(.25f));
       auto w3 = tg->g_aa_interface.GetProp(tg->g_walking_man_id);
       std::shared_ptr<AA::Scene> s3 = w3.lock();
       s3->SetStencil(true);
@@ -598,8 +605,8 @@ public:
 
     // Load and Test Models to make sure thy reuse the same VAO if loading from the same file
     {
-      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(-15, -7.5, -25));
-      tg->g_untextured_cube_id[1] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(15, -7.5, -25));
+      tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(-15, -7.5, -25));
+      tg->g_untextured_cube_id[1] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(15, -7.5, -25));
 
       // check that they are the same vaos
       auto w1 = tg->g_aa_interface.GetProp(tg->g_untextured_cube_id[0]);
@@ -617,13 +624,13 @@ public:
 
     // Load and Test Animated Models to make sure thy reuse the same VAO if loading from the same file
     {
-      tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->fullzombie_.c_str(), true, glm::vec3(-20, -30, -45), glm::vec3(.15f));
-      tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->fullzombie_.c_str(), tg->g_zombie_id[0]);
-      tg->g_aa_interface.SetAnimationOnAnimProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
+      tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->zombie_runtime_dir_path.c_str(), true, glm::vec3(-20, -30, -45), glm::vec3(.15f));
+      tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->zombie_runtime_dir_path.c_str(), tg->g_zombie_id[0]);
+      tg->g_aa_interface.SetAnimationOnProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
 
-      tg->g_zombie_id[1] = tg->g_aa_interface.AddProp(tg->fullzombie_.c_str(), true, glm::vec3(20, -30, -45), glm::vec3(.15f));
-      tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->fullzombie_.c_str(), tg->g_zombie_id[1]);
-      tg->g_aa_interface.SetAnimationOnAnimProp(tg->g_punching_anim_id, tg->g_zombie_id[1]);
+      tg->g_zombie_id[1] = tg->g_aa_interface.AddProp(tg->zombie_runtime_dir_path.c_str(), true, glm::vec3(20, -30, -45), glm::vec3(.15f));
+      tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->zombie_runtime_dir_path.c_str(), tg->g_zombie_id[1]);
+      tg->g_aa_interface.SetAnimationOnProp(tg->g_punching_anim_id, tg->g_zombie_id[1]);
 
       auto w1 = tg->g_aa_interface.GetProp(tg->g_zombie_id[0]);
       std::shared_ptr<AA::Scene> s1 = w1.lock();
@@ -705,24 +712,26 @@ public:
       glm::vec3(*tg->point_light_spec));
 
     // ground
-    for (int i = -1; i < 2; i++)
-      for (int j = -1; j < 2; j++)
-        tg->g_aa_interface.AddProp(tg->fullgroundplane.c_str(), false, glm::vec3(i * 400, -1, j * 400), glm::vec3(1.f));
-
+    {
+      std::vector<unsigned int> ground_plane_ids{};  // keep them for a moment to debug
+      for (int i = -1; i < 2; i++)
+        for (int j = -1; j < 2; j++)
+          ground_plane_ids.push_back(tg->g_aa_interface.AddProp(tg->ground_plane_runtime_dir_path.c_str(), false, glm::vec3(i * 400, -1, j * 400), glm::vec3(1.f)));
+    }
     // untextured cube
     const int CUBE_SIZE = 20;
-    tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->fullcubepath.c_str(), false, glm::vec3(0, CUBE_SIZE, 0), glm::vec3(CUBE_SIZE));
+    tg->g_untextured_cube_id[0] = tg->g_aa_interface.AddProp(tg->cube_runtime_dir.c_str(), false, glm::vec3(0, CUBE_SIZE, 0), glm::vec3(CUBE_SIZE));
 
     // peasant man
-    tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->fullpeasant_man.c_str(), false, glm::vec3(0, 0, -200), glm::vec3(1.f));
+    tg->g_peasant_man_id = tg->g_aa_interface.AddProp(tg->peasant_man_runtime_dir_path.c_str(), false, glm::vec3(0, 0, -200), glm::vec3(1.f));
 
     // man 
-    tg->g_walking_man_id = tg->g_aa_interface.AddProp(tg->fullwalking_man.c_str(), false, glm::vec3(180, 0, -100), glm::vec3(1.f));
+    tg->g_walking_man_id = tg->g_aa_interface.AddProp(tg->peasant_girl_runtime_dir_path.c_str(), false, glm::vec3(180, 0, -100), glm::vec3(1.f));
 
     // zombie with punching anim
-    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->fullzombie_.c_str(), true, glm::vec3(-180, 0, -100), glm::vec3(1.f));
-    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->fullzombie_.c_str(), tg->g_zombie_id[0]);
-    tg->g_aa_interface.SetAnimationOnAnimProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
+    tg->g_zombie_id[0] = tg->g_aa_interface.AddProp(tg->zombie_runtime_dir_path.c_str(), true, glm::vec3(-180, 0, -100), glm::vec3(1.f));
+    tg->g_punching_anim_id = tg->g_aa_interface.AddAnimation(tg->zombie_runtime_dir_path.c_str(), tg->g_zombie_id[0]);
+    tg->g_aa_interface.SetAnimationOnProp(tg->g_punching_anim_id, tg->g_zombie_id[0]);
 
 
 
@@ -860,8 +869,8 @@ public:
       std::shared_ptr<AA::Camera> local_camera_ref = tg->g_camera_ref.lock();
       local_camera_ref->SetKeepCameraToWindowSize(true);
       local_camera_ref->SetFOV(*tg->cam_fov);
-      local_camera_ref->SetPosition(glm::vec3(0, 0, 0));
-      local_camera_ref->SetPitch(-10.f);
+      local_camera_ref->SetPosition(glm::vec3(0, 10, 0));
+      local_camera_ref->SetPitch(-13.f);
       setup_fpp_fly(tg->g_cam_id, tg->g_aa_interface);
       tg->g_aa_interface.AddToOnQuit([]() { turn_off_fly(); });
 
@@ -879,11 +888,10 @@ public:
         glm::vec3(*tg->spot_light_spec)
       );
 
-
       // a point light
-      tg->point_light_loc[0] = 46.75f;
-      tg->point_light_loc[1] = 46.75f;
-      tg->point_light_loc[2] = -41.0f;
+      tg->point_light_loc[0] = 58.7328f;
+      tg->point_light_loc[1] = 83.7934f;
+      tg->point_light_loc[2] = -72.3714f;
 
       tg->g_plight1_id = tg->g_aa_interface.AddPointLight(
         glm::vec3(tg->point_light_loc[0], tg->point_light_loc[1], tg->point_light_loc[2]),
@@ -898,10 +906,11 @@ public:
 
       // fireplace room
       {
-        auto id = tg->g_aa_interface.AddProp("../../RuntimeFiles/3dmodels/fireplace_room.obj", false, glm::vec3(-10, -10, 0), glm::vec3(20));
-        //auto fr_weak = tg->g_aa_interface.GetProp(id);
-        //auto fr_strong = fr_weak.lock();
-        //fr_strong->
+        auto id = tg->g_aa_interface.AddProp(
+          tg->fireplace_room_runtime_dir_path.c_str(),
+          false,
+          glm::vec3(-20, -20, 0),
+          glm::vec3(40));
       }
     }
