+++ vcpkg.json
@@ -4,7 +4,6 @@
 	"dependencies": [
 		"glm",
 		"assimp",
-		"entt",
 		"glfw3",
 		"glad",
 		"openal-soft",
